refactor: 특수문자 구분자 처리 추가 및 관련 테스트 케이스 보강

+++ src/main/java/calculator/Application.java
@@ -21,7 +21,7 @@ public class Application {
 
             long result = calculator.add(input);
 
-            IOUtils.printResult((int) result);
+            IOUtils.printResult(result);
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             throw new IllegalArgumentException();
